feat(mllama): weight loader + dispatcher wiring — the arch is reachable
Mllama was absent from `dispatcher.rs` entirely: `loader.rs` imported
`MllamaForConditionalGeneration` and never used it, so the only live
consumer was the text-only LoRA path. Nothing could load or run it.

Adds `load_mllama_weights`, a hand-written loader — the generic one
can't express either of Mllama's two quirks. The patch-embedding
convolution needs PyTorch's `[O, I, kh, kw]` transposed to MLX's
`[O, kh, kw, I]`, and the text decoder sits under a
`language_model.model.` prefix. Current `transformers` nests both towers
under a `MllamaModel` called `model`, while the released weights predate
that, so the loader probes for whichever prefix the checkpoint actually
uses instead of guessing.

Wires `ModelArchitecture::Mllama` / `DynamicModel::Mllama` through every
match: detection (before the generic `llama` arms, which
`MllamaForConditionalGeneration` and `mllama_text_model` would otherwise
be swallowed by), load, forward, forward_hidden, forward_with_cache,
create_cache, quantize_fp8, vocab_size, hidden_size, eval, the two
DFlash target matches, and the Python arch enum.

The uniform `forward(input_ids, mask)` signature carries no images, so
those arms run **text-only** — cross-attention layers are skipped
exactly as the reference does for text-only inputs. Image conditioning
goes through `as_mllama_mut()` +
`prepare_cross_attention(..)`/`forward_full(..)`, which is also the shape
a generation loop wants: the vision features are computed once and reused
per step. We recompute their k/v projections each step where the
reference caches them; that is a perf gap, not a correctness one.

Consolidates `LoadReport` into `architectures/utils.rs` alongside the
`load_{param,optional_param,linear,layer_norm}` helpers. There were
already two byte-identical copies (gemma4, dflash_draft); both now
re-export it, so adding a third for Mllama was the alternative.

+++ crates/pmetal-bridge/src/compat/nn.rs
@@ -271,7 +271,7 @@ mod tests {
 
     /// `0.5·x·(1 + erf(x/√2))` at f64 precision, from
     /// `0.5*x*(1+math.erf(x/math.sqrt(2)))`.
-    const GELU_ERF_TABLE: [(f32, f32); 10] = [
+    const GELU_ERF_TABLE: [(f32, f64); 10] = [
         (-4.0, -0.000_126_685),
         (-2.0, -0.045_500_264),
         (-1.0, -0.158_655_254),
@@ -294,7 +294,7 @@ mod tests {
 
         for (i, &(x, want)) in GELU_ERF_TABLE.iter().enumerate() {
             assert!(
-                (got[i] - want).abs() < 1e-6,
+                (got[i] as f64 - want).abs() < 1e-6,
                 "gelu_erf({x}) = {}, want {want}",
                 got[i]
             );
@@ -314,7 +314,7 @@ mod tests {
         let x = Array::from_f32_slice(&xs, &[n as i32]);
 
         for (name, mut approx, min_gap) in [
-            ("gelu (sigmoid)", gelu(&x), 1e-2_f32),
+            ("gelu (sigmoid)", gelu(&x), 1e-2_f64),
             ("gelu_tanh_approximate", gelu_tanh_approximate(&x), 1e-4),
         ] {
             approx.eval();
@@ -322,8 +322,8 @@ mod tests {
             let worst = GELU_ERF_TABLE
                 .iter()
                 .zip(&approx)
-                .map(|(&(_, want), &got)| (got - want).abs())
-                .fold(0.0_f32, f32::max);
+                .map(|(&(_, want), &got)| (got as f64 - want).abs())
+                .fold(0.0_f64, f64::max);
             assert!(
                 worst > min_gap,
                 "{name} is within {worst:.3e} of exact gelu — closer than the \
